Keep document controls out of Nodi and verify optional package migration
